add to repo

+++ cli_lib/android/iface.go
@@ -266,6 +266,6 @@ func IsValidNinjaAddr(addr string) bool {
 	return true
 }
 
-func IconIndex(id string,mod int) int64 {
-	return int64(utils.ID2IconIdx(id,mod))
+func IconIndex(id string,mod int) int32 {
+	return int32(utils.ID2IconIdx(id,mod))
 }
